+++ src/Front/src/app/app.component.ts
@@ -10,6 +10,7 @@ import {
 import { NavigationEnd, Router, RouterOutlet } from '@angular/router';
 import { FooterComponent } from './core/layouts/footer/footer.component';
 import { NavigationComponent } from './core/layouts/navigation/navigation.component';
+import { SidebarComponent } from './core/layouts/sidebar/sidebar.component';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { filter, map, startWith } from 'rxjs';
 import { TranslateModule } from '@ngx-translate/core';
@@ -20,7 +21,7 @@ import { environment } from '../environments/environment';
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, NavigationComponent, FooterComponent, TranslateModule],
+  imports: [RouterOutlet, NavigationComponent, FooterComponent, SidebarComponent, TranslateModule],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
   schemas: [CUSTOM_ELEMENTS_SCHEMA],